@@ -39,6 +39,7 @@
     "core-js": "~3.2.1",
     "firebase-tools": "^7.13.1",
     "hammerjs": "~2.0.8",
+    "ng-lazyload-image": "^7.1.0",
     "ngrx-store-localstorage": "^8.0.0",
     "rxjs": "~6.5.2",
     "tslib": "~1.10.0",
@@ -46,8 +47,8 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.803.4",
-    "@angular/cli": "^8.3.8",
-    "@angular/compiler-cli": "~8.2.5",
+    "@angular/cli": "^8.3.26",
+    "@angular/compiler-cli": "^8.2.14",
     "@angular/language-service": "~8.2.5",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
@@ -67,6 +68,6 @@
     "ts-node": "~7.0.1",
     "tslint": "~5.15.0",
     "tslint-config-prettier": "~1.18.0",
-    "typescript": "~3.5.3"
+    "typescript": "^3.6.5"
   }
 }
