@@ -3,7 +3,7 @@ import { Entry } from 'contentful';
 export abstract class AbstractCardComponent {
   protected constructor(readonly name: string,
                         readonly index?: string,
-                        readonly object?: Entry<any>,
+                        readonly object?: any,
                         readonly routerLink?: string,
                         readonly iconClass?: string,
                         readonly col?: string,
