@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
-import UserState from 'src/app/services/store/user/user.state';
 import * as UserActions from 'src/app/services/store/user/user.action';
+import * as CartActions from 'src/app/services/store/cart/cart.action';
 import { NavigationService } from 'src/app/services/navigation/navigation.service';
+import { MatIconRegistry } from '@angular/material';
+import { DomSanitizer } from '@angular/platform-browser';
 
 @Component({
   selector: 'app-profile-page',
@@ -14,18 +16,25 @@ export class ProfileComponent implements OnInit {
   
   constructor(private router: Router,
               private store: Store<{}>,
-              private navService : NavigationService) {
-
+              private navService : NavigationService,
+              private matIconRegistry: MatIconRegistry,
+              private domSanitizer: DomSanitizer) {
+                this.matIconRegistry.addSvgIcon(
+                  'doo-approved',
+                  this.domSanitizer.bypassSecurityTrustResourceUrl("../assets/approved-green.svg")
+                );
   }
 
   ngOnInit() {
     this.navService.resetStack(['account/profile']);
     this.store.dispatch(UserActions.BeginGetUserAddressInfoAction());
     this.store.dispatch(UserActions.BeginGetFavoritesAction());
+    this.store.dispatch(UserActions.BeginGetOrdersAction());
   }
 
   public async logout() {
     this.store.dispatch(UserActions.BeginUserLogoutAction());
+    this.store.dispatch(CartActions.BeginResetCartAction());
     this.router.navigate(['']);
   }
 
