@@ -1,8 +1,9 @@
+import { UserAddressInfo } from '../user/user.model';
 
 export default class Cart {
   items?: Array<CartItem>;
   cartid? : any;
-  
+  shipping? : UserAddressInfo;
 }
 
 export class CartItem {
